requirements.txt
Django==5.2.7
django-crispy-forms==2.1
crispy-bootstrap4==2024.1
django-ckeditor==6.7.0
django-cleanup==8.1.0
Pillow==10.2.0
